+++ packages/dex-ui/src/components/Account.tsx
@@ -69,9 +69,6 @@ const Account = () => {
 							value={walletContext.currentNetwork}
 							onChange={async (ev) => await switchNetwork(ev.target.value as Networks)}
 						>
-							{walletContext.currentNetwork !== zekoTestnet && walletContext.currentNetwork !== minaTestnet && (
-								<option>N/A</option>
-							)}
 							<option value={minaMainnet}>Mina</option>
 							<option value={zekoTestnet}>Zeko</option>
 							<option value={minaTestnet}>Devnet</option>